BrainHelpers: extract placement utilities for specialised brains
Refactor HarmonicBrain to use shared helpers from a new
BrainHelpers.ts module. Future per-faction brains can re-use:
- placeAtBestCoverage(ctx, type) — highest path-coverage cell
- placeInBuffZone(ctx, type, anchors) — within range of anchors
- placeAtMaxStack(ctx, type, auras) — densest aura overlap
- bestUpgradeInBuffZone(ctx, candidates, auras) — upgrade target
  ranked by aura coverage, level as tiebreaker

Behaviour-preserving — HarmonicBrain post-refactor still wins
89-93% on harmonic|normal across three holdout seed ranges.

Negative results worth recording (not committed as code):

  MechanicalBrain (built, dropped). Tried 2 build orders:
    (1) wall-heavy opening (4 walls → 2 turrets → mortar/flame/tesla)
        — avgWave 9.5, much WORSE than BalancedBrain default 16.8.
        Walls eat early-game DPS budget, leaks pile up before any
        turret is on the board.
    (2) DPS-first (3 turrets → flame/tesla/mortar → upgrades)
        — avgWave 16.1, matches BalancedBrain default but doesn't
        improve. Upgrade-priority after wave 8 didn't move it.

  Conclusion: mechanical doesn't have a unique mechanic the way
  harmonic does (range-based aura stacking). Generic brains DO
  model conventional TD play; they just hit the same wall a
  hand-crafted brain hits, because mechanical's wall is
  late-game raw-damage shortfall, not strategy fit. Likely needs
  balance change (Turret base damage / Mortar cost) or true new
  decision primitives (sell-and-rebuild) rather than another
  hand-crafted brain.

Helpers stay because they'll be re-used by future specialised
brains where the mechanic *does* genuinely differ from generic
play (psionic status timing, possibly future per-faction work).
